% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHostParams.R
\name{checkGCC}
\alias{checkGCC}
\title{Checks the local host for GCC version in the path}
\usage{
checkGCC(OS.type = .Platform$OS.type)
}
\arguments{
\item{OS.type}{Character specifying operating system type. Defaults to \code{.Platform$OS.type}.}
}
\value{
\code{TRUE} if GCC check is successful, otherwise \code{FALSE}.
}
\description{
Performs operating system dependent check for availability of GCC.
}
\examples{
  checkGCC()

}
