% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHostParams.R
\name{checkInstallDir}
\alias{checkInstallDir}
\title{Checks the given directory for the files required for NLME run}
\usage{
checkInstallDir(installDir)
}
\arguments{
\item{installDir}{directory Location of NLME executables as set in \code{INSTALLDIR} environment variable.}
}
\value{
\code{TRUE} if all checks are successful, otherwise \code{FALSE}.
}
\description{
Checks the given directory for the files required for NLME run
}
\examples{
\dontrun{
checkInstallDir(Sys.getenv("INSTALLDIR"))
}

}
