% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHostParams.R
\name{checkMPISettings}
\alias{checkMPISettings}
\title{Check MPI settings for the given local host}
\usage{
checkMPISettings(obj)
}
\arguments{
\item{obj}{NLME Parallel Host to be checked}
}
\value{
\code{TRUE} if MPI executables are ready for running,
 otherwise \code{FALSE}. If host does not have MPI in parallel method,
 it also returns \code{TRUE}.
}
\description{
Checks if MPI settings are provided and feasible.
Check is done for the hosts where MPI parallel method is used.
}
\examples{
\dontrun{
checkMPISettings(host)
}

}
