% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHostParams.R
\name{checkRootDir}
\alias{checkRootDir}
\title{Check NLME ROOT DIRECTORY for the given local host}
\usage{
checkRootDir(obj)
}
\arguments{
\item{obj}{NLME Parallel Host to be checked}
}
\value{
\code{TRUE} if NLME ROOT DIRECTORY exists and accessible for writing,
 otherwise \code{FALSE}.
}
\description{
Checks if NLME ROOT DIRECTORY is provided and ready for writing.
That directory is used for temporary folders writing.
}
\examples{
\dontrun{
checkRootDir(host)
}

}
