% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.r
\name{downloadFileWithCat}
\alias{downloadFileWithCat}
\title{downloadFileWithCat : reads progress file and returns its content as stdout of a 'cat' command}
\usage{
downloadFileWithCat(obj, filename)
}
\arguments{
\item{obj}{job handle to an NLME job}

\item{filename}{file name related to remote directory}
}
\description{
downloadFileWithCat : reads progress file and returns its content as stdout of a 'cat' command
}
