% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_params.r
\docType{data}
\name{ERR_CENTRAL_DIFF}
\alias{ERR_CENTRAL_DIFF}
\title{central}
\format{An object of class \code{numeric} of length 1.}
\usage{
ERR_CENTRAL_DIFF
}
\description{
central
}
\keyword{datasets}
