% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_params.r
\docType{data}
\name{ERR_FORWARD_DIFF}
\alias{ERR_FORWARD_DIFF}
\title{forward diff}
\format{An object of class \code{numeric} of length 1.}
\usage{
ERR_FORWARD_DIFF
}
\description{
forward diff
}
\keyword{datasets}
