% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.r
\name{fitmodel}
\alias{fitmodel}
\title{fitmodel() : Method to execute an NLME simple estimation}
\usage{
fitmodel(hostPlatform, dataset, params = NULL, model = NULL,
  runInBackground = TRUE)
}
\arguments{
\item{hostPlatform}{How to execute the run(NlmeParallelHost)}

\item{dataset}{Dataset and model information(NlmeDataset)}

\item{params}{Engine parameters(NlmeEngineExtraParams)}

\item{model}{Optional PK/PD model}

\item{runInBackground}{TRUE will run in background and return prompt(Bool)}
}
\description{
fitmodel() : Method to execute an NLME simple estimation
}
\examples{

input=read.csv("PKPD_Data.csv")

model = pkmodel(numComp=1,
               absorption = Extravascular,
               modelName="PkModel")

dataset = defaultDataset(model,input)

initColMapping(model)= input

modelColumnMapping(model)=c(id="ID", CObs="Conc",Aa="Dose")

initFixedEffects(model)=c(tvV=16,tvCl=41,tvV2=7,tvCl2=14)

host = NlmeParallelHost(sharedDirectory=Sys.getenv("NLME_ROOT_DIRECTORY"),
                       parallelMethod=NlmeParallelMethod("LOCAL_MPI"),
                       hostName="MPI",
                       numCores=4)

params = NlmeEngineExtraParams(PARAMS_METHOD=METHOD_FOCE_LB,
                             PARAMS_NUM_ITERATIONS=1000)

job = fitmodel(host,dataset,params,model)

}
