% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_params.r
\docType{data}
\name{METHOD_LAPLACIAN}
\alias{METHOD_LAPLACIAN}
\title{LAPLACIAN}
\format{An object of class \code{numeric} of length 1.}
\usage{
METHOD_LAPLACIAN
}
\description{
LAPLACIAN
}
\keyword{datasets}
