% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_params.r
\docType{data}
\name{METHOD_NAIVE_POOLED}
\alias{METHOD_NAIVE_POOLED}
\title{NAIVE_POOLED}
\format{An object of class \code{numeric} of length 1.}
\usage{
METHOD_NAIVE_POOLED
}
\description{
NAIVE_POOLED
}
\keyword{datasets}
