% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\docType{class}
\name{NlmeBootstrapParams-class}
\alias{NlmeBootstrapParams-class}
\alias{NlmeBootstrapParams}
\title{NlmeBootstrapParams : Represents parameters for a bootstrap run}
\arguments{
\item{numReplicates}{Number of bootstrap replicates to run}

\item{initialEstimates}{Run and initial estimation to get initial values(1|0)}

\item{numRetries}{Number of times to retry a failed replicated}

\item{randomNumSeed}{Seed for random number generator}

\item{confidenceLevel}{Confidence Level}

\item{stratifyColumns}{What columns to stratify on(comma separated)}
}
\description{
NlmeBootstrapParams : Represents parameters for a bootstrap run
}
\examples{

boot = NlmeBootstrapParams(numReplicates=50,
                           randomNumSeed=1234)

}
