% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_search.r
\docType{class}
\name{NlmeCovariateEffectModel-class}
\alias{NlmeCovariateEffectModel-class}
\alias{NlmeCovariateEffectModel}
\title{NlmeCovariateEffectModel : NLME covariate effects model object class}
\arguments{
\item{numCovariates}{Number of covariate effects}

\item{covariateList}{Comma separted list of covariate effects names}

\item{scenarioNames}{Comma separted list of scenario names}

\item{isDefault}{Comma separted list of flags}

\item{degreesOfFreedom}{Comma separted list of degress of freedome}
}
\description{
Class represents an NLME covariate effects model
}
\examples{
NlmeCovariateEffectModel(...)

}
\keyword{NLME,}
\keyword{NlmeCovariateEffectModel}
