% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_params.r
\docType{class}
\name{NlmeEngineExtraParams-class}
\alias{NlmeEngineExtraParams-class}
\alias{NlmeEngineExtraParams}
\title{NlmeEngineExtraParams : Defines all extra engine parameters}
\arguments{
\item{PARAMS_IS_POPULATION}{is this a population model}

\item{PARAMS_METHOD}{what engine to use
METHOD_QRPEM|METHOD_IT2S_EM|METHOD_FOCE_LB|METHOD_FIRST_ORDER|
METHOD_LAPLACIAN|METHOD_NAIVE_POOLED}

\item{PARAMS_NUM_ITERATIONS}{maximum number of iterations}

\item{PARAMS_ODE}{What ODE to use
common : ODE_STIFF=2|ODE_NON_STIFF=3|ODE_AUTO_DETECT=5|MATRIX_EXP=6
others : ODE_LSODE=1|ODE_LSODE_ANA_JAC=2|ODE_RK=3|ODE_LSODA=4|
         ODE_LSODA_ANA_JAC=5|MATRIX_EXP=6}

\item{PARAMS_ALLOW_ANAL_GRAD}{n  0, or 1 to allow analytic gradients (default 0)}

\item{PARAMS_NON_PARAMETRIC}{n  number of nonparametric generations (or 0)}

\item{PARAMS_RESTART}{n     0, or 1 to restart}

\item{PARAMS_NORDERAGQ}{n   number of AGQ points per axis (or 0)}

\item{PARAMS_FOCEHESE}{n    0 for numerical hessian, 1 for foce}

\item{PARAMS_VERBOSE}{n     0 or 1}

\item{PARAMS_STDERR}{n    0=ERR_NONE, 1=ERR_CENTRAL_DIFF, 2=ERR_FORWARD_DIFF}

\item{PARAMS_LAMETH}{n    LAGL Method 1, 2, or 3}

\item{PARAMS_LAGL_PRECISION}{n      LAGL nDigit}

\item{PARAMS_LA_STDERR}{n       LAGL tolerance}

\item{PARAMS_BLUP_PREC}{n      BLUP Method 1, 2, or 3}

\item{PARAMS_BLUP_PREC}{n      BLUP nDigit}

\item{PARAMS_BLUP_STDERR}{n       BLUP tolerance}

\item{PARAMS_PCWRES_NREP}{n  number of reps for PCWRES simulation}

\item{PARAMS_ISAMPLE}{n     number of samples for QRPEM}

\item{PARAMS_MAPASSIST}{n   0 (default) or >0, to enable map assist and specify periodicity}

\item{PARAMS_MAPNP}{n       map naive-pool control}

\item{PARAMS_IMPSAMPDOF}{n  importance sampling control (-3 to 30) 
PARAMS_IMP_SAMPLE_NORMAL=0
PARAMS_IMP_SAMPLE_DBL_EXP=1
PARAMS_IMP_SAMPLE_DIRECT=2
PARAMS_IMP_SAMPLE_MIXTURE_2=-2
PARAMS_IMP_SAMPLE_MIXTURE_3=-3}

\item{PARAMS_MCPEM}{n       0 for QRPEM, 1 for MCPEM}

\item{PARAMS_PRESAMPLE}{n   0 (default) for no presampling}

\item{PARAMS_SIR_SAMP}{n     number of SIR samples}

\item{PARAMS_BURNIN}{n      number of burn-in samples}

\item{PARAMS_NONOMEGA_BURN}{n   0 (default) or 1, to use non-omega burn-in}

\item{PARAMS_NO_BURN}{n   0 (default) or 1, to suppress burn-in}

\item{PARAMS_START_FROM_POST}{n   0 (default) or 1, to start from saved posteriors}

\item{PARAMS_ACC_RATIO}{n    acceptance ratio}

\item{PARAMS_SCRAMBLE}{n    0 for None, 1 for Owen, 2 for Faure-Tezuka}

\item{PARAMS_PAR_DERIV_NUM_STEPS}{n      partial derivative # steps}

\item{PARAMS_PAR_DERIV_DELTA}{n      partial derivative delta}

\item{PARAMS_AUTO_LOG_TRANSFORM}{n      0, or 1(default) to enable log-translate of data}
}
\description{
NlmeEngineExtraParams : Defines all extra engine parameters
}
\examples{
 
           param=NlmeEngineExtraParams(PARAMS_METHOD=METHOD_FOCE_LB,
                                       PARAMS_NUM_ITERATIONS=1000)
           param=NlmeEngineExtraParams(PARAMS_METHOD=METHOD_QRPEM,
                                       PARAMS_NUM_ITERATIONS=300)
           param=NlmeEngineExtraParams(PARAMS_METHOD=METHOD_QRPEM,
                                       PARAMS_NUM_ITERATIONS=300,
                                       PARAMS_IS_POPULATION=TRUE,
                                       PARAM_ODE=ODE_STIFF)

}
