% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.r
\name{NlmeJobStatus}
\alias{NlmeJobStatus}
\title{NlmeJobStatus : reads progress file and returns the status of a job}
\usage{
NlmeJobStatus(obj)
}
\arguments{
\item{job}{handle to an NLME job}
}
\description{
NlmeJobStatus : reads progress file and returns the status of a job
}
\examples{
    status=NlmeJobStatus(jobHandle)

}
