% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.r
\docType{class}
\name{NlmeObservationVar-class}
\alias{NlmeObservationVar-class}
\alias{NlmeObservationVar}
\title{NlmeObservationVar : Describes an observation(observe,multi,...)}
\arguments{
\item{name}{of observation variable}

\item{type}{of observation}

\item{xaxis}{One of:VPC_XAXIS_T,VPC_XAXIS_TAD,VPC_XAXIS_PRED,VPC_XAXIS_OTHER}

\item{binningMethod}{VPC_BIN_NONE,VPC_BIN_KMEANS,VPC_BIN_EXP_CENTERS,VPC_BIN_EXP_BOUNDARIES}

\item{binningOption}{comma separated list to specify centers or boundary values}

\item{quantilesValues}{comma separated list}

\item{quantilesSecondaryValues}{comma separated list}
}
\description{
NlmeObservationVar : Describes an observation(observe,multi,...)
}
\examples{

var = NlmeObservationVar(
                      name="Cobs",
                      type=VPC_OBSERVE_T,
                      xaxis=VPC_XAXIS_TAD,
                      binningMethod=VPC_BIN_NONE,
                      quantilesValues ="5,50,95")

}
