% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_param.r
\docType{class}
\name{NlmeParallelHost-class}
\alias{NlmeParallelHost-class}
\alias{NlmeParallelHost}
\title{NlmeParallelHost}
\arguments{
\item{sharedDirectory}{This is the directory where the run happens in}

\item{installationDirectory}{Directory containing NLME libraries/scripts}

\item{hostName}{visual name of the host(default local)}

\item{machineName}{IP address or name of the host(default local)}

\item{hostType}{Windows or Linux}

\item{numCores}{Number of compute cores}

\item{parallelMethod}{None|Multicore|SGE|SGE_MPI|TORQUE|TORQUE_MPI
MPI|LSF|LSF_MPI|LOCAL_MPI}

\item{userAuthentication}{user credential for remote system}

\item{scriptPath}{Optional initialization remote script path}

\item{isLocal}{Is this a local or remote host}
}
\description{
NLME Parallel Host object class
}
\details{
Class represents an NLME parallel host
It can either be local or remote
}
\examples{
NlmeParallelHost(...)

}
\keyword{NLME,}
\keyword{NlmeParallelHost}
