% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_estimation.r
\docType{class}
\name{NlmeProfileParameters-class}
\alias{NlmeProfileParameters-class}
\alias{NlmeProfileParameters}
\title{NLME Profile variable}
\arguments{
\item{How}{to apply profile variables. USE_DELTA|USE_PERCENTAGE}

\item{list}{of profile variables}
}
\description{
Class represents an NLME profile pertubation variable
}
\examples{
profile1 = NlmeProfileVar("tvV","9.548","-2,-1,0,1,2")
profile2 = NlmeProfileVar("tvCl","3.219","-1,0,1")
profiles= NlmeProfileParameters("USE_DELTA",c(profile1,profile2))

}
