% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_estimation.r
\docType{class}
\name{NlmeProfileVar-class}
\alias{NlmeProfileVar-class}
\alias{NlmeProfileVar}
\title{NLME Profile variable}
\arguments{
\item{fixed}{effect name}

\item{initial}{value}

\item{list}{of values to pertubate by(either delta or percentage)}
}
\description{
Class represents an NLME profile pertubation variable
}
\examples{
NlmeProfileVar("tvV",9.95,"-2,-1,0,1,2")

}
