% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_param.r
\docType{class}
\name{NlmeRemoteExecutor-class}
\alias{NlmeRemoteExecutor-class}
\alias{NlmeRemoteExecutor}
\title{NlmeRemoteExecutor}
\arguments{
\item{sharedDirectory}{This is the directory where the run happens in}

\item{installationDirectory}{Directory containing NLME libraries/scripts}

\item{hostName}{IP or name of remote host}

\item{userAuthentication}{credential for user to log into remote system}
}
\description{
Remote Executor
}
