% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortby_estimation.r
\docType{class}
\name{NlmeScenario-class}
\alias{NlmeScenario-class}
\alias{NlmeScenario}
\title{NLME scenario object}
\arguments{
\item{scenarioName}{Name of the scenario}

\item{covariatesList}{Comma separated Indexs of covariates effects to use for this scenario}
}
\description{
Class represents an NLME Scenario
}
\examples{

scenario1=NlmeScenario("SC0001","1")
    Use covariate effect index 1
scenario2=NlmeScenario("SC0002","")
    Do not use any covariate effects
scenario3=NlmeScenario("SC0002","2")
    Use second covariate effect in the model
scenario4=NlmeScenario("SC0002","1,2")
   Use 1st and 2nd covariate effect


}
