% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.r
\docType{class}
\name{NlmeSimTableDef-class}
\alias{NlmeSimTableDef-class}
\alias{NlmeSimTableDef}
\title{NlmeSimTableDef : Parameters for VPC/Simulation runs}
\arguments{
\item{name}{Name of the generated simulation file}

\item{timesList}{List of time values}

\item{variablesList}{List of variables}

\item{timeAfterDose}{Time after dose flag}
}
\description{
NlmeSimTableDef : Parameters for VPC/Simulation runs
}
