% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.r
\docType{class}
\name{NlmeSimulationParams-class}
\alias{NlmeSimulationParams-class}
\alias{NlmeSimulationParams}
\title{NlmeSimulationParams : Parameters for simulation runs}
\arguments{
\item{numReplicates}{Number of replicates to simulate}

\item{seed}{Random number generator seed}

\item{simulationTables}{(NlmeSimTableDef)}

\item{isPopulation}{Simulating a population model(default=TRUE).  The rest of arguments applies to individual models only}

\item{numPoints}{Number of points in simulation}

\item{maxXRange}{Max value of independent variable}

\item{yVariables}{comma separated list of Y variables}

\item{simAtObs}{Simulate values at observed values of ivar}
}
\description{
NlmeSimulationParams : Parameters for simulation runs
}
\examples{

table1=NlmeSimTableDef(name="simulate.csv",timesList="0,2,4,12,24",
                       variablesList="V,Cl",
                       timeAfterDose=TRUE)

simParam = NlmeSimulationParams(numReplicates=10,
                      seed=29423,
                      simulationTables = c(table1))

simParam = NlmeSimulationParams(isPopulation=FALSE,
                      numPoints=100,
                      maxXRange=50,
                      yVariables="C,A1",
                      simulationTables = c(table1))

}
