% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortby_estimation.r
\docType{class}
\name{NlmeSortColumns-class}
\alias{NlmeSortColumns-class}
\alias{NlmeSortColumns}
\title{NLME Sort column object}
\arguments{
\item{sortColumnList}{Comma separted list of covariate effects names}
}
\description{
Class represents an NLME Sort columns
}
\examples{
NlmeSortColumns("Country,City")

}
