% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_search.r
\docType{class}
\name{NlmeStepwiseParams-class}
\alias{NlmeStepwiseParams-class}
\alias{NlmeStepwiseParams}
\title{NlmeStepwiseParams : NLME stepwise parameters}
\arguments{
\item{addPValue}{Threshold for adding a covariate effect}

\item{removePValue}{Threshold for removing a covariate effect}

\item{method}{-2LL|AIC|BIC}
}
\description{
Class represents an NLME Stepwise search parameters
}
\examples{
 NlmeStepwiseParams(addPValue,removePValue,method)

}
