% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_param.r
\docType{class}
\name{NlmeUserAuthentication-class}
\alias{NlmeUserAuthentication-class}
\alias{NlmeUserAuthentication}
\title{NlmeUserAuthentication}
\arguments{
\item{userName}{How the user is identified to the remote system}

\item{privateKeyFile}{Ssh generated private key}

\item{userPassword}{Password to login into remote system}
}
\description{
Authentication records
}
