% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.r
\docType{class}
\name{NlmeVpcParams-class}
\alias{NlmeVpcParams-class}
\alias{NlmeVpcParams}
\title{NlmeVpcParams : Parameters for VPC runs}
\arguments{
\item{numReplicates}{Number of replicates to simulate}

\item{seed}{Random number generator seed}

\item{predCorrection}{One of VPC_PRED_NONE,VPC_PRED_PROPOTIONAL,VPC_PRED_ADDITIVE}

\item{predVarCorr}{flag to use Prediction Variance Correction}

\item{stratifyColumns}{List of covariates for Stratified PC}

\item{observactionVars}{(NlmeObservationVar)}

\item{simulationTables}{Optional list of simulatio tables (NlmeSimTableDef)}
}
\description{
NlmeVpcParams : Parameters for VPC runs
}
\examples{

observe1 = NlmeObservationVar(name="Cobs",
                             type=VPC_OBSERVE_T,
                             xaxis=VPC_XAXIS_TAD,
                             binningMethod=VPC_BIN_NONE,
                             quantilesValues ="5,50,95")

observe2 = NlmeObservationVar(name="Iobs",
                             type=VPC_MULTI_T,
                             xaxis=VPC_XAXIS_PRED,
                             quantilesValues ="5,50,95")

observe3 = NlmeObservationVar(name="Eobs",
                             type=VPC_LL_T,
                             timeToEvent="seq(1,10)"
                             quantilesValues ="5,50,95")

  table1=NlmeSimTableDef(name="simulate.csv",
                         timesList="0,2,4,12,24",
                         variablesList="V,Cl",
                         timeAfterDose=TRUE)

vpc = NlmeVpcParams(numReplicates=10,
                      seed=29423,
                      predCorrection=VPC_PRED_PROPOTIONAL,
                      predVarCorr=TRUE,
                      stratifyColumns="sex,race,dosing",
                      observationVars=c(observe1,observe2,observe3),
                      simulationTables=c(table1))

}
