% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performBootstrap.R
\name{performBootstrap}
\alias{performBootstrap}
\title{NLME Bootstrap Function}
\usage{
performBootstrap(args, allowIntermediateResults = TRUE, reportProgress = FALSE)
}
\arguments{
\item{args}{Arguments for bootstrap execution}

\item{allowIntermediateResults}{Set to \code{TRUE} to return intermediate results}

\item{reportProgress}{Set to \code{TRUE} to report progress}
}
\value{
Directory path where NLME job was executed
}
\description{
Runs an NLME bootstrap job in parallel and produces summaries
}
\keyword{Bootstrap}
\keyword{NLME}
