% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performEstimationOnSortColumns.R
\name{performEstimationOnSortColumns}
\alias{performEstimationOnSortColumns}
\title{Sort specification for multiple estimations}
\usage{
performEstimationOnSortColumns(args, reportProgress = FALSE)
}
\arguments{
\item{args}{a vector of arguments provided as the following:
c(method, install_directory, shared_directory, localWorkingDir,
 nlmeArgsFile, numColumns, ColumnNames, NumProc, workflowName)}

\item{reportProgress}{whether it is required to report the progress
(for local jobs usually)}
}
\value{
Directory path where NLME job was executed
}
\description{
Runs multiple estimations sorting the input dataset by
requested columns and creating multiple data sets
}
\keyword{NLME}
