% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performParallelNLMERun.R
\name{performParallelNLMERun}
\alias{performParallelNLMERun}
\title{Runs a set of NLME jobs in parallel}
\usage{
performParallelNLMERun(
  args,
  partialJob = FALSE,
  allowIntermediateResults = TRUE,
  progressStage = "",
  func = "",
  func_arg = NULL,
  reportProgress = FALSE
)
}
\arguments{
\item{args}{a vector of arguments provided as the following:
c(jobType, parallelMethod, install_dir, shared_directory,
localWorkingDir, controlFile, NumProc, workflow_name, fixefUnits)}

\item{partialJob}{is \code{TRUE} if it is not required to stop the job
as for covariate stepwise search}

\item{allowIntermediateResults}{is \code{TRUE} if intermediate results
are possible like for sorting}

\item{progressStage}{stage of analysis to be reported}

\item{func}{function to be executed after NLME job}

\item{func_arg}{arguments to be provided to the function by name provided above}

\item{reportProgress}{whether it is required to report the progress
(for local jobs usually)}
}
\value{
Directory path where NLME job was executed
}
\description{
Runs a set of NLME jobs in parallel
}
