% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performProfileEstimation.R
\name{performProfileEstimation}
\alias{performProfileEstimation}
\title{NLME a profile estimation run on list of fixed effects}
\usage{
performProfileEstimation(args, reportProgress = FALSE)
}
\arguments{
\item{args}{Arguments for profile estimation}

\item{reportProgress}{Set to \code{TRUE} to report progress}
}
\value{
Directory path where NLME job was executed
}
\description{
This function runs multiple estimations sorting the input dataset by
requested columns and creating multiple data sets
Runs are also generated for all profiling variables
}
\keyword{NLME}
