% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performShotgunCovarSearch.R
\name{performShotgunCovarSearch}
\alias{performShotgunCovarSearch}
\title{Shotgun covariate search}
\usage{
performShotgunCovarSearch(args, reportProgress = FALSE)
}
\arguments{
\item{args}{a vector of arguments provided as the following:
c(jobType, parallelMethod, install_dir, shared_directory,
localWorkingDir, controlFile, NumProc, workflow_name, fixefUnits)}

\item{reportProgress}{whether it is required to report the progress
(for local jobs usually)}
}
\value{
Directory path where NLME job was executed
}
\description{
Runs a set of possible covariate sets in parallel
}
\keyword{NLME}
\keyword{ShotgunCovariateSearch}
