% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.r
\name{print.SimpleNlmeJob}
\alias{print.SimpleNlmeJob}
\title{print : reads progress file and prints out its contents}
\usage{
\method{print}{SimpleNlmeJob}(obj)
}
\arguments{
\item{job}{handle to an NLME job}
}
\description{
print : reads progress file and prints out its contents
}
\examples{
    print(jobHandle)

}
