% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_estimation.r
\name{profilePertubate}
\alias{profilePertubate}
\title{profilePertubate() : Method to execute an NLME profile pertubation}
\usage{
profilePertubate(hostPlatform, params = NULL, profiles, model = NULL,
  sortColumns = NlmeSortColumns(""), scenarios = list(),
  runInBackground = TRUE)
}
\arguments{
\item{hostPlatform}{How to execute the run(NlmeParallelHost)}

\item{params}{Engine parameters(NlmeEngineExtraParams)}

\item{profiles}{profiles to pertubate(NlmeProfileParameters)}

\item{model}{PK/PD model}

\item{sortColumns}{Optional list of columns to sort and fit(NlmeSortColumns)}

\item{scenarios}{Optional list of scenarios to fit(NlmeScenario)}

\item{runInBackground}{TRUE will run in background and return prompt(Bool)}
}
\description{
profilePertubate() : Method to execute an NLME profile pertubation
}
\examples{


param = NlmeEngineExtraParams(PARAMS_METHOD=METHOD_FOCE_LB,
                             PARAMS_NUM_ITERATIONS=1000)

profile1 = NlmeProfileVar("tvV",9.548,"-2,-1,0,1,2")
profile2 = NlmeProfileVar("tvCl",0.919,"-0.5,0,1.5")
profiles= NlmeProfileParameters("USE_DELTA",c(profile1,profile2))

job = profilePertubate(defaultHost,params,profiles,model)

}
