% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconnectToBootstrapNLMERun.R
\name{reconnectToBootstrapNLMERun}
\alias{reconnectToBootstrapNLMERun}
\title{Use to reconnect to a grid job}
\usage{
reconnectToBootstrapNLMERun(args)
}
\arguments{
\item{args}{Arguments for reconnecting to bootstrap grid run}
}
\value{
Directory path where NLME job was executed
}
\description{
Use to reconnect to a grid job
}
\keyword{Bootstrap}
\keyword{NLME}
