% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{RunBootstrap}
\alias{RunBootstrap}
\title{RunBootstrap() : Method to execute an NLME Bootstrap}
\usage{
RunBootstrap(hostPlatform, dataset, params, bootParams,
  runInBackground = TRUE, workingDir = NULL)
}
\arguments{
\item{hostPlatform}{How to execute the run(NlmeParallelHost)}

\item{dataset}{Dataset and model information(NlmeDataset)}

\item{params}{Engine parameters(NlmeEngineExtraParams)}

\item{bootParams}{Bootstrap parameters(NlmeBootstrapParams)}

\item{runInBackground}{TRUE will run in background and return prompt(Bool)}

\item{workingDir}{where to run the job}
}
\description{
RunBootstrap() : Method to execute an NLME Bootstrap
}
\examples{

dataset = NlmeDataset()

bootParams = NlmeBootstrapParams(numReplicates=50,randomNumSeed=1234)

param = NlmeEngineExtraParams(PARAMS_METHOD=METHOD_FOCE_LB,
                             PARAMS_NUM_ITERATIONS=1000)
bootParam = NlmeBootstrapParams(numReplicates=50, randomNumSeed=1234)

job = RunBootstrap(defaultHost,dataset,params,bootParams)

while (!(NlmeJobStatus(job) == "Finished" || NlmeJobStatus(job) == "Failed") )
{
    print(NlmeJobStatus(job))
    print(job)
    Sys.sleep(5)
}

}
