% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_search.r
\name{RunShotgunSearch}
\alias{RunShotgunSearch}
\title{RunShotgunSearch : Method to execute an NLME shotgun covariate search}
\usage{
RunShotgunSearch(hostPlatform, dataset, params, covariateModel,
  runInBackground = TRUE, workingDir = NULL)
}
\arguments{
\item{hostPlatform}{How to execute the run(NlmeParallelHost)}

\item{dataset}{Dataset and model information(NlmeDataset)}

\item{params}{Engine parameters(NlmeEngineExtraParams)}

\item{covariateModel}{Covariate Effects Model(NlmeCovariateEffectModel)}

\item{runInBackground}{TRUE will run in background and return prompt(Bool)}

\item{workingDir}{where to run the job}
}
\description{
RunShotgunSearch : Method to execute an NLME shotgun covariate search
}
\examples{

dataset = NlmeDataset()

param = NlmeEngineExtraParams(PARAMS_METHOD=METHOD_FOCE_LB,
                             PARAMS_NUM_ITERATIONS=1000)
covariateModel = ReadNlmeCovariateEffectModel("covariates.txt")

job = RunShotgunSearch(defaultHost,dataset,params,covariateModel)

}
