% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.r
\name{RunSimpleEstimation}
\alias{RunSimpleEstimation}
\title{RunSimpleEstimation() : Method to execute an NLME simple estimation}
\usage{
RunSimpleEstimation(hostPlatform, dataset, params,
  runInBackground = TRUE)
}
\arguments{
\item{hostPlatform}{How to execute the run(NlmeParallelHost)}

\item{dataset}{Dataset and model information(NlmeDataset)}

\item{params}{Engine parameters(NlmeEngineExtraParams)}

\item{runInBackground}{TRUE will run in background and return prompt(Bool)}
}
\description{
RunSimpleEstimation() : Method to execute an NLME simple estimation
}
\examples{

dataset = NlmeDataset()

param = NlmeEngineExtraParams(PARAMS_METHOD=METHOD_FOCE_LB,
                             PARAMS_NUM_ITERATIONS=1000)

job = RunSimpleEstimation(defaultHost,dataset,params)

}
