% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortby_estimation.r
\name{RunSortByEstimation}
\alias{RunSortByEstimation}
\title{RunSortByEstimation() : Method to execute multiple NLME estimation w sort key}
\usage{
RunSortByEstimation(hostPlatform, dataset, params, sortColumns,
  scenarios = list(), runInBackground = TRUE, workingDir = NULL)
}
\arguments{
\item{hostPlatform}{How to execute the run(NlmeParallelHost)}

\item{dataset}{Dataset and model information(NlmeDataset)}

\item{params}{Engine parameters(NlmeEngineExtraParams)}

\item{runInBackground}{TRUE will run in background and return prompt(Bool)}

\item{workingDir}{where to run the job}

\item{sortList}{list of sort columns (NlmeSortColumns)}

\item{scenariosList}{list of scenario names(c())}
}
\description{
RunSortByEstimation() : Method to execute multiple NLME estimation w sort key
}
\examples{

dataset = NlmeDataset()

param = NlmeEngineExtraParams(PARAMS_METHOD=METHOD_FOCE_LB,
                             PARAMS_NUM_ITERATIONS=1000)
listCovariateEffectNames(model)

scenario1=NlmeScenario("SC0001","1")
scenario2=NlmeScenario("SC0002","")
scenario3=NlmeScenario("SC0002","2")
scenario4=NlmeScenario("SC0002","1,2")
sortColumns=NlmeSortColumns("group,sex")
scenariosList = c(scenario1,scenario2,scenario3,secnario4)

job = RunSortByEstimation(defaultHost,dataset,params,sortColumns,scenariosList,runInBackground)
}
