% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_search.r
\name{RunStepwiseSearch}
\alias{RunStepwiseSearch}
\title{RunStepwiseSearch() : Method to execute an NLME stepwise covariate search}
\usage{
RunStepwiseSearch(hostPlatform, dataset, params, covariateModel,
  stepwiseParams, runInBackground = TRUE, workingDir)
}
\arguments{
\item{hostPlatform}{How to execute the run(NlmeParallelHost)}

\item{dataset}{Dataset and model information(NlmeDataset)}

\item{params}{Engine parameters(NlmeEngineExtraParams)}

\item{covariateModel}{Covariate Effects Model(NlmeCovariateEffectModel)}

\item{runInBackground}{TRUE will run in background and return prompt(Bool)}

\item{workingDir}{where to run the search}

\item{stepwiseParam}{stepwise parameters(NlmeStepwiseParams)}
}
\description{
RunStepwiseSearch() : Method to execute an NLME stepwise covariate search
}
\examples{

dataset = NlmeDataset()

param = NlmeEngineExtraParams(PARAMS_METHOD=METHOD_FOCE_LB,
                             PARAMS_NUM_ITERATIONS=1000)
covariateModel = ReadNlmeCovariateEffectModel("covariates.txt")

stepwiseParam = NlmeStepwiseParams(0,01, 0.001, "-2LL")

job = RunStepwiseSearch(defaultHost,dataset,params,covariateModel,stepwiseParam)

}
