% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.r
\name{RunVpcSimulation}
\alias{RunVpcSimulation}
\title{RunVpcSimulation() : Method to execute an NLME VPC simulation}
\usage{
RunVpcSimulation(hostPlatform, dataset, params, vpcParams = NULL,
  simParams = NULL, runInBackground = TRUE)
}
\arguments{
\item{hostPlatform}{How to execute the run(NlmeParallelHost)}

\item{dataset}{Dataset and model information(NlmeDataset)}

\item{params}{Engine parameters(NlmeEngineExtraParams)}

\item{vpcParams}{VPC parameters(NlmeVpcParams)}

\item{runInBackground}{TRUE will run in background and return prompt(Bool)}
}
\description{
RunVpcSimulation() : Method to execute an NLME VPC simulation
}
\examples{

dataset = NlmeDataset()

vpcParams = NlmeVpcParams()

param = NlmeEngineExtraParams(PARAMS_METHOD=METHOD_FOCE_LB,
                             PARAMS_NUM_ITERATIONS=1000)

job = RunVpcSimulation(defaultHost,dataset,params,vpcParams,simParams)

}
