% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_search.r
\name{shotgunSearch}
\alias{shotgunSearch}
\title{shotgunSearch()}
\usage{
shotgunSearch(hostPlatform, params = NULL, covariateModel, model,
  runInBackground = TRUE)
}
\arguments{
\item{hostPlatform}{How to execute the run(NlmeParallelHost)}

\item{params}{Engine parameters(NlmeEngineExtraParams)}

\item{covariateModel}{Covariate Effects Model(NlmeCovariateEffectModel)}

\item{model}{PK/PD model}

\item{runInBackground}{TRUE will run in background and return prompt(Bool)}

\item{shotgunParam}{shotgun parameters(NlmeStepwiseParams)}
}
\description{
Method to execute an NLME shotgun covariate search
}
\examples{


param = NlmeEngineExtraParams(PARAMS_METHOD=METHOD_FOCE_LB,
                             PARAMS_NUM_ITERATIONS=1000)

job = shotgunSearch(defaultHost,
                    params,
                    covariateModel(model),
                    model)

}
