% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.r
\name{simmodel}
\alias{simmodel}
\title{simmodel}
\usage{
simmodel(hostPlatform, dataset, simParams = NULL, model = NULL,
  runInBackground = TRUE)
}
\arguments{
\item{hostPlatform}{How to execute the run(NlmeParallelHost)}

\item{dataset}{Dataset and model information(NlmeDataset)}

\item{simParams}{Simulation parameters(NlmeSimulationParam)}

\item{model}{optional PK/PD model}

\item{runInBackground}{TRUE will run in background and return prompt(Bool)}
}
\description{
Method to execute an NLME simulation
}
\examples{


SimTableObs = NlmeSimTableDef(name = "SimTableObs.csv",
                                 timesList = "0,1,2,4,4.9,55.1,56,57,59,60",
                                 variablesList = "C, CObs",
                                 timeAfterDose = FALSE)

simParams = NlmeSimulationParams(numReplicates = 50,
                            seed = 3527,
                            simulationTables = c(SimTableObs))

job = simmodel(defaultHost,dataset,simParams,model)

}
