% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortby_estimation.r
\name{sortfit}
\alias{sortfit}
\title{sortfit() : Method to execute an NLME simple estimation w sort keys}
\usage{
sortfit(hostPlatform, params = NULL, sortColumns, scenarios = list(),
  model, runInBackground = TRUE)
}
\arguments{
\item{hostPlatform}{How to execute the run(NlmeParallelHost)}

\item{params}{Engine parameters(NlmeEngineExtraParams)}

\item{model}{PK/PD model}

\item{runInBackground}{TRUE will run in background and return prompt(Bool)}

\item{sortList}{list of sort columns (NlmeSortColumns)}

\item{scenariosList}{list of scenario names(c())}
}
\description{
sortfit() : Method to execute an NLME simple estimation w sort keys
}
\examples{

param = NlmeEngineExtraParams(PARAMS_METHOD=METHOD_FOCE_LB,
                             PARAMS_NUM_ITERATIONS=1000)
listCovariateEffectNames(model)

scenario1=NlmeScenario("SC0001","1")
scenario2=NlmeScenario("SC0002","")
scenario3=NlmeScenario("SC0002","2")
scenario4=NlmeScenario("SC0002","1,2")
scenarios = c(scenario1,scenario2,scenario3,scenario4)

sortColumns=NlmeSortColumns("group,sex")

job = sortfit(defaultHost,params,sortColumns,scenarios,model)

}
