% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_search.r
\name{stepwiseSearch}
\alias{stepwiseSearch}
\title{stepwiseSearch()}
\usage{
stepwiseSearch(hostPlatform, dataset, params = NULL, covariateModel,
  stepwiseParams, model = NULL, runInBackground = TRUE)
}
\arguments{
\item{hostPlatform}{How to execute the run(NlmeParallelHost)}

\item{dataset}{Dataset and model information(NlmeDataset)}

\item{params}{Engine parameters(NlmeEngineExtraParams)}

\item{covariateModel}{Covariate Effects Model(NlmeCovariateEffectModel)}

\item{model}{Optional PK/PD model}

\item{runInBackground}{TRUE will run in background and return prompt(Bool)}

\item{stepwiseParam}{stepwise parameters(NlmeStepwiseParams)}
}
\description{
Method to execute an NLME stepwise covariate search
}
\examples{

dataset = NlmeDataset()

param = NlmeEngineExtraParams(PARAMS_METHOD=METHOD_FOCE_LB,
                             PARAMS_NUM_ITERATIONS=1000)

sp = NlmeStepwiseParams(0.01, 0.001, "-2LL")

job=stepwiseSearch(host,
                   dataset,
                   params,
                   covariateModel(covarModel),
                   sp,
                   model)

}
