% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utl.r
\name{summarizeBootstrap}
\alias{summarizeBootstrap}
\title{Reads bootstrap file and transposes the data
It computes the following statistics per sample run:
Mean, Stderr, CV, Median, Percentiles}
\usage{
summarizeBootstrap(localWorkingDir, confidenceLevel)
}
\arguments{
\item{local}{working directory}

\item{Confidence}{level}
}
\description{
It will generate ascii files : BootTheta.txt and BootThetaStacked.txt
}
\examples{
summarizeBootstrap(localWorkingDir,confidenceLevel)
}
\keyword{summarizeBootstrap}
