% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.r
\name{bootstrap}
\alias{bootstrap}
\title{bootstrap() : Method to execute an NLME Bootstrap}
\usage{
bootstrap(hostPlatform, params, bootParams, model,
  runInBackground = TRUE)
}
\arguments{
\item{hostPlatform}{How to execute the run(NlmeParallelHost)}

\item{params}{Engine parameters(NlmeEngineExtraParams)}

\item{bootParams}{Bootstrap parameters(NlmeBootstrapParams)}

\item{model}{PK/PD model}

\item{runInBackground}{TRUE will run in background and return prompt(Bool)}
}
\description{
bootstrap() : Method to execute an NLME Bootstrap
}
\examples{

params = NlmeEngineExtraParams(PARAMS_METHOD=METHOD_FOCE_LB,
                             PARAMS_NUM_ITERATIONS=1000)
bootParams = NlmeBootstrapParams(numReplicates=5, randomNumSeed=1234)

job = bootstrap(defaultHost,params,bootParams,model)

while (!(NlmeJobStatus(job) == "Finished" || NlmeJobStatus(job) == "Failed") )
{
    print(NlmeJobStatus(job))
    print(job)
    Sys.sleep(5)
}

}
