% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utl.r
\name{getTableNames}
\alias{getTableNames}
\title{Table names from the column definition file}
\usage{
getTableNames(columnDefinitionFilename, simtbl = FALSE)
}
\arguments{
\item{columnDefinitionFilename}{path to NLME column definition file}

\item{simtbl}{logical. \code{TRUE} extracts simulation tables,
\code{FALSE} extracts simple tables.}
}
\value{
vector of names of the tables in column definition file if any,
empty string otherwise
}
\description{
Extracts table names from the column definition file
}
\examples{
\dontrun{
getTableNames("cols1.txt", simtbl = TRUE)
}

}
