% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utl.r
\name{performParallelNLMERun}
\alias{performParallelNLMERun}
\title{NLME Generic run Function}
\usage{
performParallelNLMERun(args, partialJob = FALSE,
  allowIntermediateResults = TRUE, progressStage = "", func = "",
  func_arg = NULL, reportProgress = FALSE)
}
\arguments{
\item{JobType}{COVAR_SEARCH, GENERIC,}

\item{ParalleMethod}{How to run, MPI, Multicore, Grid}

\item{Install_Directory}{Directory where NLME libraries, scripts are}

\item{Shared_Directory}{Directory accessible by all nodes}

\item{Local_Working_Directory}{Where input files are and results will be stored}

\item{Control}{file File containg instructions how to run}

\item{Number_of_Cores_to_use}{Number of cores to use for multicore runs
Control file format :
Line1 : name of pml model file
Line2 : Comma separated list of input files to copy to remote disk
Line3 : Comma separated list of outpu files to grab from NLME build processes
Line4 : Number of NLME jobs to run
Line5..LineN : comma separated parameters for each run
      JobName/ScenarioName
      Main NLME run control file(i.e. nlmeargs.txt)
      Extra arguments to pass to NLME7.exe ( optional )     
      Filename to check to declare the run a success(i.e. out.txt )
      Optional Space sepaated List of out fiels to return from each run
      Optional Space sepaated List of out files to return from each run
      ( These will be tagged with the jobname )}
}
\description{
This function runs a set of  NLME jobs in parallel
}
\examples{
performParallelNLMERun(method,install_directory, shared_directory localWorkingDir ControlFile NumProc )

}
\keyword{Bootstrap}
\keyword{NLME,}
