% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utl.r
\name{performProfileEstimation}
\alias{performProfileEstimation}
\title{NLME a profile estimation run on list of fixed effects}
\usage{
performProfileEstimation(args, reportProgress = FALSE)
}
\arguments{
\item{ParalleMethod}{How to run, MPI, Multicore, Grid}

\item{Install_Directory}{Directory where NLME libraries, scripts are}

\item{Shared_Directory}{Directory accessible by all nodes}

\item{Local_Working_Directory}{Where input files are and results will be stored}

\item{NlmeArgsFile}{Arguments to NLME executable}

\item{NumColmns}{Number of sort columns}

\item{ColumnNames}{Array of column names}

\item{fixedEffectList}{Array of fixed effects to pertubate}

\item{NumOfCores}{Number of compute nodes to use}

\item{WorkflowName}{Name of workflow to use in summary report}
}
\description{
This function runs multiple estimations sorting the input dataset by 
requested columns and creating multiple data sets
Runs are also generated for all profiling variables
}
\examples{
performProfileEstimation(method, install_directory, shared_directory localWorkingDir modelFile nlmeArgsFile numColumns ColumnNames fixedEffectList NumProc, workflowName )


}
\keyword{NLME,}
