% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utl.r
\name{performShotgunCovarSearch}
\alias{performShotgunCovarSearch}
\title{NLME shotgun covariate search}
\usage{
performShotgunCovarSearch(args, reportProgress = FALSE)
}
\arguments{
\item{ParalleMethod}{How to run, MPI, Multicore, Grid}

\item{Install_Directory}{Directory where NLME libraries, scripts are}

\item{Shared_Directory}{Directory accessible by all nodes}

\item{Local_Working_Directory}{Where input files are and results will be stored}

\item{ModelFile}{Filename of the model}

\item{NlmeArgsFile}{Arguments to NLME executable}

\item{InputFiles}{List of input files}

\item{Num_Covariates}{Number of covariates}

\item{Covariate_Names}{Array of covariate names}

\item{Num_Of_Cores}{Number of compute nodes to use}

\item{Workflow_Name}{Name of workflow to use in summary report}
}
\description{
This function runs a shotgun covariate NLME job in parallel
}
\examples{
performShotgunCovarSearch(method, install_directory, shared_directory localWorkingDir modelFile nlmeArgsFile numCovariates CovariateNames NumProc )


}
\keyword{NLME,}
\keyword{ShotgunCovariateSearch}
