% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utl.r
\name{performStepwiseCovarSearch}
\alias{performStepwiseCovarSearch}
\title{NLME stepwise covariate search}
\usage{
performStepwiseCovarSearch(args, reportProgress = FALSE)
}
\arguments{
\item{ParalleMethod}{How to run, MPI, Multicore, Grid}

\item{Install_Directory}{Directory where NLME libraries, scripts are}

\item{Shared_Directory}{Directory accessible by all nodes}

\item{Local_Working_Directory}{Where input files are and results will be stored}

\item{ModelFile}{Filename of the model}

\item{NlmeArgsFile}{Arguments to NLME executable}

\item{InputFiles}{List of input files}

\item{NumCovariates}{Number of covariates}

\item{CovariateNames}{Array of covariate names}

\item{Criteria}{What criteria to use for convergance}

\item{AddPValue}{Threashold value for adding variables}

\item{RemovePValue}{Threashold value for removing variables}

\item{NumOfCores}{Number of compute nodes to use}

\item{WorkflowName}{Name of workflow to use in summary report}
}
\description{
This function runs a shotgun covariate NLME job in parallel
}
\examples{
performStepwiseCovarSearch(method, install_directory, shared_directory localWorkingDir modelFile nlmeArgsFile numCovariates CovariateNames Criteria addPValue removePValue NumProc )


}
\keyword{NLME,}
\keyword{StepwiseCovariateSearch}
