% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utl.r
\name{reconnectToBootstrapNLMERun}
\alias{reconnectToBootstrapNLMERun}
\title{NLME Bootstrap Function}
\usage{
reconnectToBootstrapNLMERun(args)
}
\arguments{
\item{ParalleMethod}{How to run, MPI, Multicore, Grid}

\item{Install_Directory}{Directory where NLME libraries, scripts are}

\item{Shared_Directory}{Directory accessible by all nodes}

\item{Local_Working_Directory}{Where input files are and results will be stored}

\item{Optimization_Engine}{Which Pharsight NLME engine to use}

\item{Num_Iterations}{Maximum iterations}

\item{Num_Samples}{Number of bootstrap samples to run}

\item{Max_Tries}{How many times to try a sample that does not converge}

\item{Model_Filename}{PML model}

\item{Column_Definition_Filename}{Mapping of columns}

\item{Data_File}{NLME data file}

\item{Start_Seed}{Initial seed}

\item{Extra_Arguments_File}{Other parameters to pass to engines}

\item{List_of_Files_to_copy_to_shared_directory}{List of files to copy to the remote systems}

\item{Number_of_Cores_to_use}{Number of cores to use for multicore runs}

\item{Confidence_Level}{Confidence level to summarize for}
}
\description{
This function reconnects to a grid job
}
\examples{
reconnectToBootstrapNLMERun(method,install_directory, shared_directory localWorkingDir engine num_iterations num_samples max_tries model_file column_def_file data_file start_seed extra_args_file files_to_copy NumProc ConfidenceLevel gridDirerctory)

}
\keyword{Bootstrap}
\keyword{NLME,}
