% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.r
\name{vpcmodel}
\alias{vpcmodel}
\title{vpcmodel}
\usage{
vpcmodel(hostPlatform, vpcParams = NULL, model, runInBackground = TRUE)
}
\arguments{
\item{hostPlatform}{How to execute the run(NlmeParallelHost)}

\item{vpcParams}{VPC parameters(NlmeVpcParam)}

\item{model}{PK/PD model}

\item{runInBackground}{TRUE will run in background and return prompt(Bool)}
}
\description{
Method to execute an NLME visual predictive check
}
\examples{

obsVars = GetObservationVariables(model@dataset)

observationParameters(obsVars[[1]])=c(xaxis=VPC_XAXIS_T,
                                      binningMethod=VPC_BIN_NONE,
                                      quantilesValues ="5,50,95")

vpcParams = NlmeVpcParams(numReplicates=2,
                       seed=1234,
                       observationVars=obsVars)


job = vpcmodel(defaultHost,vpcParams,model)

}
